refactor: extract audible live-probe tests into probe.rs, deepen provider module cohesion
Move 4 #[ignore] live-probe tests and their exclusive helpers into
a dedicated probe.rs module. Move shared probe-summary helpers
(LibraryProbeSummary, library_probe_summary, etc.) to module level
with #[cfg(test)] gating. Keep all cancellation/session cleanup/purge
truth in session_lifecycle.rs. No provider-private leaks; public
strip stays command-only; Supplemental PDF policy remains explicit.

Reduces audible/mod.rs from 1954 to 1672 LOC.
